refactor: migrate tests from MemoryKvBackend
